+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * IoC now supports default param values for non-optional injections:
     If no binding present for a param is bound but the param has a default value
     the default value will be used for injection.
+  * implemented #117: injection stack on binding exceptions
   * improved performance of annotation parsing
   * improved annotation cache storage functions api